Adding Volume cloning files
