+++ tests/functional-tests/tests/remotecluster_volume_provisioning.py
@@ -2832,181 +2832,7 @@ def test_driver_volume_expansion_3():
     value_pod = [{"mount_path": "/usr/share/nginx/html/scale", "read_only": "False"}]
     driver_object.test_dynamic(value_sc, value_pvc_passed=value_pvc, value_pod_passed=value_pod)
 
-
-def test_driver_volume_cloning_pass_1():
-    value_sc = {"volBackendFs": data["remoteFs"], "clusterId": data["remoteid"]}
-    value_pvc = [{"access_modes": "ReadWriteMany", "storage": "1Gi"}]
-    value_clone_passed = {"clone_pvc": [{"access_modes": "ReadWriteMany", "storage": "1Gi"}, {
-        "access_modes": "ReadWriteOnce", "storage": "1Gi"}]}
-    driver_object.test_dynamic(value_sc, value_pvc_passed=value_pvc,
-                               value_clone_passed=value_clone_passed)
-
-
-def test_driver_volume_cloning_pass_2():
-    value_sc = {"volBackendFs": data["remoteFs"], "filesetType": "dependent",
-                "clusterId": data["remoteid"]}
-    value_pvc = [{"access_modes": "ReadWriteMany", "storage": "1Gi"}]
-    value_clone_passed = {"clone_pvc": [{"access_modes": "ReadWriteMany", "storage": "1Gi"}, {
-        "access_modes": "ReadWriteOnce", "storage": "1Gi"}]}
-    driver_object.test_dynamic(value_sc, value_pvc_passed=value_pvc,
-                               value_clone_passed=value_clone_passed)
-
-
-def test_driver_volume_cloning_pass_3():
-    value_sc = {"volBackendFs": data["remoteFs"], "clusterId": data["remoteid"]}
-    value_pvc = [{"access_modes": "ReadWriteMany", "storage": "1Gi"}]
-    value_clone_passed = {"clone_pvc": [{"access_modes": "ReadWriteMany", "storage": "11Gi"}, {
-        "access_modes": "ReadWriteOnce", "storage": "8Gi"}]}
-    driver_object.test_dynamic(value_sc, value_pvc_passed=value_pvc,
-                               value_clone_passed=value_clone_passed)
-
-
-def test_driver_volume_cloning_multiple_clones():
-    value_sc = {"volBackendFs": data["remoteFs"], "clusterId": data["remoteid"]}
-    value_pvc = [{"access_modes": "ReadWriteMany", "storage": "1Gi"}]
-    value_clone_passed = {"clone_pvc": [{"access_modes": "ReadWriteMany", "storage": "1Gi"}],
-                          "number_of_clones": 5}
-    driver_object.test_dynamic(value_sc, value_pvc_passed=value_pvc,
-                               value_clone_passed=value_clone_passed)
-
-
-def test_driver_volume_cloning_chain_clones():
-    value_sc = {"volBackendFs": data["remoteFs"], "clusterId": data["remoteid"]}
-    value_pvc = [{"access_modes": "ReadWriteMany", "storage": "1Gi"}]
-    value_clone_passed = {"clone_pvc": [{"access_modes": "ReadWriteMany", "storage": "1Gi"}],
-                          "clone_chain": 1}
-    driver_object.test_dynamic(value_sc, value_pvc_passed=value_pvc,
-                               value_clone_passed=value_clone_passed)
-
-
-def test_driver_volume_cloning_expand_before():
-    value_sc = {"volBackendFs": data["remoteFs"],
-                "clusterId": data["remoteid"], "allow_volume_expansion": True}
-    value_pvc = [{"access_modes": "ReadWriteMany",
-                  "storage": "9Gi", "volume_expansion_storage": ["11Gi"]}]
-    value_clone_passed = {"clone_pvc": [{"access_modes": "ReadWriteMany", "storage": "11Gi"},
-                                        {"access_modes": "ReadWriteMany", "storage": "9Gi", "reason": "new PVC request must be greater than or equal in size"}]}
-    driver_object.test_dynamic(value_sc, value_pvc_passed=value_pvc,
-                               value_clone_passed=value_clone_passed)
-
-
-def test_driver_volume_cloning_with_subpath():
-    value_sc = {"volBackendFs": data["remoteFs"], "clusterId": data["remoteid"], "permissions": "777",
-                "gid": data["r_gid_number"], "uid": data["r_uid_number"]}
-    value_pvc = [{"access_modes": "ReadWriteMany", "storage": "1Gi"}]
-    value_clone_passed = {"clone_pvc": [{"access_modes": "ReadWriteMany", "storage": "1Gi"}, {
-        "access_modes": "ReadWriteOnce", "storage": "1Gi"}]}
-    value_pod = [{"mount_path": "/usr/share/nginx/html/scale", "read_only": "False",
-                  "sub_path": ["sub_path_mnt", "sub_path_mnt_2", "sub_path_mnt3"], "volumemount_readonly":[False, False, True], "runAsUser": "2000", "runAsGroup": "5000"}]
-    driver_object.test_dynamic(value_sc, value_pvc_passed=value_pvc,
-                               value_pod_passed=value_pod, value_clone_passed=value_clone_passed)
-
-
-def test_driver_volume_cloning_fail_1():
-    value_sc = {"volBackendFs": data["remoteFs"], "clusterId": data["remoteid"]}
-    value_pvc = [{"access_modes": "ReadWriteMany", "storage": "1Gi"}]
-    value_clone_passed = {"clone_pvc": [{"access_modes": "ReadWriteMany", "storage": "1Gi", "reason": "same storage class for cloning"}],
-                          "clone_sc": {"volBackendFs": data["remoteFs"], "clusterId": data["remoteid"]}}
-    driver_object.test_dynamic(value_sc, value_pvc_passed=value_pvc,
-                               value_clone_passed=value_clone_passed)
-
-
-def test_driver_volume_cloning_fail_2():
-    value_sc = {"volBackendFs": data["remoteFs"], "clusterId": data["remoteid"]}
-    value_pvc = [{"access_modes": "ReadWriteMany", "storage": "1Gi"}]
-    value_clone_passed = {"clone_pvc": [{"access_modes": "ReadWriteMany", "storage": "1Gi", "reason": "same storage class for cloning"}],
-                          "clone_sc": {"volBackendFs": data["remoteFs"], "filesetType": "dependent",
-                                       "clusterId": data["remoteid"]}}
-    driver_object.test_dynamic(value_sc, value_pvc_passed=value_pvc,
-                               value_clone_passed=value_clone_passed)
-
-
-def test_driver_volume_cloning_fail_3():
-    value_sc = {"volBackendFs": data["remoteFs"], "clusterId": data["remoteid"]}
-    value_pvc = [{"access_modes": "ReadWriteMany", "storage": "1Gi"}]
-    value_clone_passed = {"clone_pvc": [{"access_modes": "ReadWriteMany", "storage": "1Gi", "reason": "same storage class for cloning"}],
-                          "clone_sc": {"volBackendFs": data["remoteFs"], "volDirBasePath": data["r_volDirBasePath"]}}
-    driver_object.test_dynamic(value_sc, value_pvc_passed=value_pvc,
-                               value_clone_passed=value_clone_passed)
-
-
-def test_driver_volume_cloning_fail_4():
-    value_sc = {"volBackendFs": data["remoteFs"], "filesetType": "dependent",
-                "clusterId": data["remoteid"]}
-    value_pvc = [{"access_modes": "ReadWriteMany", "storage": "1Gi"}]
-    value_clone_passed = {"clone_pvc": [{"access_modes": "ReadWriteMany", "storage": "1Gi", "reason": "same storage class for cloning"}],
-                          "clone_sc": {"volBackendFs": data["remoteFs"], "clusterId": data["remoteid"]}}
-    driver_object.test_dynamic(value_sc, value_pvc_passed=value_pvc,
-                               value_clone_passed=value_clone_passed)
-
-
-def test_driver_volume_cloning_fail_5():
-    value_sc = {"volBackendFs": data["remoteFs"], "filesetType": "dependent",
-                "clusterId": data["remoteid"]}
-    value_pvc = [{"access_modes": "ReadWriteMany", "storage": "1Gi"}]
-    value_clone_passed = {"clone_pvc": [{"access_modes": "ReadWriteMany", "storage": "1Gi", "reason": "same storage class for cloning"}],
-                          "clone_sc": {"volBackendFs": data["remoteFs"], "filesetType": "dependent",
-                                       "clusterId": data["remoteid"]}}
-    driver_object.test_dynamic(value_sc, value_pvc_passed=value_pvc,
-                               value_clone_passed=value_clone_passed)
-
-
-def test_driver_volume_cloning_fail_6():
-    value_sc = {"volBackendFs": data["remoteFs"], "filesetType": "dependent",
-                "clusterId": data["remoteid"]}
-    value_pvc = [{"access_modes": "ReadWriteMany", "storage": "1Gi"}]
-    value_clone_passed = {"clone_pvc": [{"access_modes": "ReadWriteMany", "storage": "1Gi", "reason": "same storage class for cloning"}],
-                          "clone_sc": {"volBackendFs": data["remoteFs"], "r_volDirBasePath": data["volDirBasePath"]}}
-    driver_object.test_dynamic(value_sc, value_pvc_passed=value_pvc,
-                               value_clone_passed=value_clone_passed)
-
-
-def test_driver_volume_cloning_fail_7():
-    value_sc = {"volBackendFs": data["remoteFs"], "volDirBasePath": data["r_volDirBasePath"]}
-    value_pvc = [{"access_modes": "ReadWriteMany", "storage": "1Gi"}]
-    value_clone_passed = {"clone_pvc": [{"access_modes": "ReadWriteMany", "storage": "1Gi", "reason": "same storage class for cloning"}],
-                          "clone_sc": {"volBackendFs": data["remoteFs"], "clusterId": data["remoteid"]}}
-    driver_object.test_dynamic(value_sc, value_pvc_passed=value_pvc,
-                               value_clone_passed=value_clone_passed)
-
-
-def test_driver_volume_cloning_fail_8():
-    value_sc = {"volBackendFs": data["remoteFs"], "volDirBasePath": data["r_volDirBasePath"]}
-    value_pvc = [{"access_modes": "ReadWriteMany", "storage": "1Gi"}]
-    value_clone_passed = {"clone_pvc": [{"access_modes": "ReadWriteMany", "storage": "1Gi", "reason": "same storage class for cloning"}],
-                          "clone_sc": {"volBackendFs": data["remoteFs"], "filesetType": "dependent",
-                                       "clusterId": data["remoteid"]}}
-    driver_object.test_dynamic(value_sc, value_pvc_passed=value_pvc,
-                               value_clone_passed=value_clone_passed)
-
-
-def test_driver_volume_cloning_fail_9():
-    value_sc = {"volBackendFs": data["remoteFs"], "volDirBasePath": data["r_volDirBasePath"]}
-    value_pvc = [{"access_modes": "ReadWriteMany", "storage": "1Gi"}]
-    value_clone_passed = {"clone_pvc": [{"access_modes": "ReadWriteMany", "storage": "1Gi", "reason": "same storage class for cloning"}],
-                          "clone_sc": {"volBackendFs": data["remoteFs"], "r_volDirBasePath": data["volDirBasePath"]}}
-    driver_object.test_dynamic(value_sc, value_pvc_passed=value_pvc,
-                               value_clone_passed=value_clone_passed)
-
-
-def test_driver_volume_cloning_fail_10():
-    value_sc = {"volBackendFs": data["remoteFs"], "clusterId": data["remoteid"]}
-    value_pvc = [{"access_modes": "ReadWriteMany", "storage": "5Gi"}]
-    value_clone_passed = {"clone_pvc": [{"access_modes": "ReadWriteMany", "storage": "1Gi",
-                                         "reason": "new PVC request must be greater than or equal in size"}]}
-    driver_object.test_dynamic(value_sc, value_pvc_passed=value_pvc,
-                               value_clone_passed=value_clone_passed)
-
-
-def test_driver_volume_cloning_fail_11():
-    value_sc = {"volBackendFs": data["remoteFs"], "volDirBasePath": data["r_volDirBasePath"]}
-    value_pvc = [{"access_modes": "ReadWriteMany", "storage": "1Gi"}]
-    value_clone_passed = {"clone_pvc": [{"access_modes": "ReadWriteMany", "storage": "1Gi", "reason": "Volume cloning for directories for remote file system is not supported"},
-                                        {"access_modes": "ReadWriteOnce", "storage": "1Gi", "reason": "Volume cloning for directories for remote file system is not supported"}]}
-    driver_object.test_dynamic(value_sc, value_pvc_passed=value_pvc,
-                               value_clone_passed=value_clone_passed)
-
-
+    
 def test_driver_compression_1():
     value_sc = {"volBackendFs": data["remoteFs"], "compression": "z", "clusterId": data["remoteid"]}
     value_pvc = [{"access_modes": "ReadWriteMany", "storage": "1Gi"}] * 2
